requirements.txt
async-timeout==4.0.2
beehive-infra==0.0.144
boto3==1.24.53
botocore==1.27.53
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cramjam==2.5.0
Deprecated==1.2.13
fastparquet==0.8.1
Flask==2.2.2
fsspec==2022.7.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
networkx==2.8.5
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pyarrow==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-louvain==0.16
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
statsd==3.3.0
Unidecode==1.3.4
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
